@@ -14,8 +14,8 @@
     "medium-editor": "^5.23.3",
     "npm": "^8.16.0",
     "quill-image-resize": "^3.0.9",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-icons": "^4.4.0",
     "react-identicons": "^1.1.7",
     "react-medium-editor": "^1.8.1",
